+++ web/packages/pyxis-components/src/public/organisms/BookingRules/BookingRules.tsx
@@ -1,5 +1,5 @@
 import { clsx } from 'clsx';
-import { pyxisPart } from '../../utils/parts';
+import { pyxisPart } from '../../../utils/parts';
 import './BookingRules.css';
 
 export type BookingRulesProps = {